refactor: remove duplicated code in back-office for graphql queries #74580
