+++ libs/shared/src/lib/utils/public-api.ts
@@ -1,2 +1,3 @@
 export * from './validators/cron.validator';
 export * from './graphql/connection.type';
+export * from './update-queries';
